requirements.txt
opencv_python==4.4.0.42
numpy==1.18.2
pandas==1.0.5
Pillow==9.0.1
